+++ src/main/java/one/talon/model/CardLedgerPointsEntryIntegrationAPI.java
@@ -34,7 +34,7 @@ import org.threeten.bp.OffsetDateTime;
 public class CardLedgerPointsEntryIntegrationAPI {
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Long id;
+  private Integer id;
 
   public static final String SERIALIZED_NAME_CREATED = "created";
   @SerializedName(SERIALIZED_NAME_CREATED)
@@ -42,7 +42,7 @@ public class CardLedgerPointsEntryIntegrationAPI {
 
   public static final String SERIALIZED_NAME_PROGRAM_ID = "programId";
   @SerializedName(SERIALIZED_NAME_PROGRAM_ID)
-  private Long programId;
+  private Integer programId;
 
   public static final String SERIALIZED_NAME_CUSTOMER_PROFILE_I_D = "customerProfileID";
   @SerializedName(SERIALIZED_NAME_CUSTOMER_PROFILE_I_D)
@@ -73,7 +73,7 @@ public class CardLedgerPointsEntryIntegrationAPI {
   private BigDecimal amount;
 
 
-  public CardLedgerPointsEntryIntegrationAPI id(Long id) {
+  public CardLedgerPointsEntryIntegrationAPI id(Integer id) {
     
     this.id = id;
     return this;
@@ -85,12 +85,12 @@ public class CardLedgerPointsEntryIntegrationAPI {
   **/
   @ApiModelProperty(example = "123", required = true, value = "ID of the transaction that adds loyalty points.")
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -117,7 +117,7 @@ public class CardLedgerPointsEntryIntegrationAPI {
   }
 
 
-  public CardLedgerPointsEntryIntegrationAPI programId(Long programId) {
+  public CardLedgerPointsEntryIntegrationAPI programId(Integer programId) {
     
     this.programId = programId;
     return this;
@@ -129,12 +129,12 @@ public class CardLedgerPointsEntryIntegrationAPI {
   **/
   @ApiModelProperty(example = "324", required = true, value = "ID of the loyalty program.")
 
-  public Long getProgramId() {
+  public Integer getProgramId() {
     return programId;
   }
 
 
-  public void setProgramId(Long programId) {
+  public void setProgramId(Integer programId) {
     this.programId = programId;
   }
